# Conflicts:
#	ethereumj-core/src/main/resources/ethereumj.conf

+++ ethereumj-core/src/main/java/org/ethereum/net/rlpx/discover/NodeManager.java
@@ -36,9 +36,7 @@ import static java.lang.Math.min;
 public class NodeManager implements Functional.Consumer<DiscoveryEvent>{
     static final org.slf4j.Logger logger = LoggerFactory.getLogger("discover");
 
-    // to avoid checking for null
-    private static NodeStatistics DUMMY_STAT = new NodeStatistics(new Node(new byte[0], "dummy.node", 0));
-    private boolean PERSIST;
+    private final boolean PERSIST;
 
     private static final long LISTENER_REFRESH_RATE = 1000;
     private static final long DB_COMMIT_RATE = 1 * 60 * 1000;
@@ -142,7 +140,7 @@ public class NodeManager implements Functional.Consumer<DiscoveryEvent>{
     private void dbRead() {
         try {
             db = mapDBFactory.createTransactionalDB("network/discovery");
-            if (config.databaseReset()) {
+            if (config.databaseReset() && config.databaseResetBlock() == 0) {
                 logger.info("Resetting DB Node statistics...");
                 db.delete("nodeStats");
             }
@@ -248,7 +246,7 @@ public class NodeManager implements Functional.Consumer<DiscoveryEvent>{
     }
 
     public NodeStatistics getNodeStatistics(Node n) {
-        return discoveryEnabled ? getNodeHandler(n).getNodeStatistics() : DUMMY_STAT;
+        return getNodeHandler(n).getNodeStatistics();
     }
 
     @Override
@@ -398,7 +396,9 @@ public class NodeManager implements Functional.Consumer<DiscoveryEvent>{
         peerConnectionManager.close();
         try {
             nodeManagerTasksTimer.cancel();
-            dbWrite();
+            if (PERSIST) {
+                dbWrite();
+            }
         } catch (Exception e) {
             logger.warn("Problems canceling nodeManagerTasksTimer", e);
         }
